Create GetExpiredUsers
This one liner will query Active Directory for expired user accounts. It selects their name, username, distinguished name, and the date of expiration. It will sort them by most outdated account and list them in the terminal. Quick and simple without parsing a CSV, but can be easily modified to do that.